renamed file, removed alexnet as an option
