+++ tests/test_pretrained_train.py
@@ -4,7 +4,7 @@ import os
 import pandas as pd
 from backend.common.dataset import dataset_from_zipped, loader_from_zipped
 from backend.common.constants import DEFAULT_TRANSFORM
-from backend.dl.pytorch_pretrained import pytorch_pretrained
+from backend.dl.pretrained import pytorch_pretrained
 
 train_dir = "" if (os.getcwd()).split("\\")[-1].split("/")[-1] == "tests" else "tests"
 backend_dir = (
